Backend/requirements.txt
torch==2.5.1+cu121
torchvision==0.20.1+cu121
numpy==2.4.3
pandas==3.0.2
scipy==1.17.1
scikit-learn==1.8.0
fastapi==0.136.0
uvicorn==0.44.0
pyyaml==6.0.3
matplotlib==3.10.8
seaborn==0.13.2
tqdm==4.67.3
httpx==0.28.1
